requirements.txt
pytilities~=1.2.0
pytest~=6.1.2
nltk~=3.5
pyenchant~=3.1.1
websockets~=8.1